+++ lib/yamload/validation.rb
@@ -1,7 +1,0 @@
-module Yamload
-  module Validation
-  end
-end
-
-require 'yamload/validation/hash'
-require 'yamload/validation/result'